[refactor] 검색 context 분리
