+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { RouterProvider } from 'react-router-dom'
 import { router } from './routes/Router'
 import { AuthProvider } from './context/AuthProvider'
-import { SearchProvider } from './context/SearchContext'
+import { SearchProvider } from './context/search/SearchProvider'
 
 function App() {
   return (
