+++ src/components/common/Header.tsx
@@ -5,20 +5,18 @@ import person from '@/assets/icon/person.svg'
 import alarm from '@/assets/icon/alarm.svg'
 import search from '@/assets/icon/search.svg'
 import { Link, NavLink, useNavigate } from 'react-router-dom'
-import { useEffect, useRef, useState } from 'react'
+import { useContext, useEffect, useRef, useState } from 'react'
 import { useAuth } from '@/hooks/auth/useAuth'
-import { useSearch } from '@/context/SearchContext'
+import { SearchContext } from '@/context/search/SearchContext'
 
-export const Header = ({
-  onSearch
-}: {
-  onSearch: (keyword: string) => void
-}) => {
+export const Header = () => {
   const [showHeader, setShowHeader] = useState(true)
   const lastScrollY = useRef(0)
   const { isAuthenticated, logout } = useAuth()
   const [searchKeyword, setSearchKeyword] = useState('')
-  const { setKeyword } = useSearch()
+
+  const { setKeyword } = useContext(SearchContext)
+
   const navigate = useNavigate()
 
   console.log('Header 렌더링됨, isAuthenticated:', isAuthenticated)
