--- src/context/SearchContext.tsx
@@ -1,24 +1,0 @@
-import { createContext, useContext, useState, ReactNode } from 'react'
-
-interface SearchContextType {
-  keyword: string
-  setKeyword: (value: string) => void
-}
-
-const SearchContext = createContext<SearchContextType | undefined>(undefined)
-
-export const SearchProvider = ({ children }: { children: ReactNode }) => {
-  const [keyword, setKeyword] = useState('')
-
-  return (
-    <SearchContext.Provider value={{ keyword, setKeyword }}>
-      {children}
-    </SearchContext.Provider>
-  )
-}
-
-export const useSearch = () => {
-  const context = useContext(SearchContext)
-  if (!context) throw new Error('useSearch must be used within a SearchProvider')
-  return context
-}
