+++ src/hooks/useSearchPortfoilo.ts
@@ -1,0 +1,8 @@
+// import { useSearch } from '@/context/SearchContext'
+// import { useContext } from 'react'
+
+// export const useSearchPortfoilo = () => {
+//   const { keyword, setKeyword } = useContext(SearchContext)
+
+//   return { keyword, setKeyword }
+// }
